# docs/openapi-doc.yml
openapi: 3.1.0
info:
    title: Firepit API
    version: 1.7.0
    description: |
        Firepit is a realtime communication platform built with Next.js 16, React 19, Bun, and Appwrite.

        This specification covers the public and first-party client API surface currently used by the application.
        Internal debug and test-only endpoints are intentionally omitted.
    contact:
        name: Firepit
        url: https://github.com/acarlson33/firepit
    license:
        name: GNU General Public License v3.0
        url: https://opensource.org/licenses/GPL-3.0
servers:
    - url: http://localhost:3000
      description: Local development server
    - url: https://firepit.example.com
      description: Example production server
...
components:
    securitySchemes:
        sessionCookie:
            type: apiKey
            in: cookie
            name: a_session_*
            description: Appwrite session cookie used by authenticated first-party clients
    schemas:
...
        Success:
            type: object
            properties:
                success:
                    type: boolean
                    default: true
...
        FileAttachment:
            type: object
            properties:
                fileId:
                    type: string
                fileName:
                    type: string
                fileSize:
                    type: integer
                fileType:
                    type: string
                fileUrl:
                    type: string
                    format: uri
                thumbnailUrl:
                    type: string
                    format: uri
        Message:
            type: object
            properties:
                $id:
                    type: string
                userId:
                    type: string
                userName:
                    type: string
                text:
                    type: string
                channelId:
                    type: string
                serverId:
                    type: string
                imageFileId:
                    type: string
                imageUrl:
                    type: string
                    format: uri
                replyToId:
                    type: string
                mentions:
                    type: array
                    items:
                        type: string
                attachments:
                    type: array
                    items:
                        $ref: "#/components/schemas/FileAttachment"
                removedAt:
                    type: string
                    format: date-time
                removedBy:
                    type: string
                $createdAt:
                    type: string
                    format: date-time
...
paths:
...
    /api/messages:
        post:
            tags: [Messages]
            summary: Create a channel message
            security:
                - sessionCookie: []
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            type: object
                            required: [channelId]
                            properties:
                                text:
                                    type: string
                                channelId:
                                    type: string
                                serverId:
                                    type: string
                                imageFileId:
                                    type: string
                                imageUrl:
                                    type: string
                                    format: uri
                                replyToId:
                                    type: string
                                mentions:
                                    type: array
                                    items:
                                        type: string
                                attachments:
                                    type: array
                                    items:
                                        $ref: "#/components/schemas/FileAttachment"
            responses:
                "200":
                    description: Created message
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    message:
                                        $ref: "#/components/schemas/Message"
        patch:
            tags: [Messages]
            summary: Edit a channel message
            security:
                - sessionCookie: []
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            type: object
                            required: [messageId]
                            properties:
                                messageId:
                                    type: string
                                text:
                                    type: string
            responses:
                "200":
                    description: Updated message
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    message:
                                        $ref: "#/components/schemas/Message"
        delete:
            tags: [Messages]
            summary: Delete a channel message
            security:
                - sessionCookie: []
            parameters:
                - name: messageId
                  in: query
                  required: true
                  schema:
                      type: string
            responses:
                "200":
                    description: Message deleted
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Success"
...